Make minor changes noted in code review
