+++ api/handlers/devicehandler.py
@@ -3,57 +3,18 @@ import datetime as dt
 from .. import base
 from .. import config
 from .. import util
-from ..dao import containerstorage, APIPermissionException, APINotFoundException
-from ..types import Origin
+from ..auth import require_drone, require_login, require_superuser
+from ..dao import containerstorage, APINotFoundException
+from ..validators import validate_data
 
 log = config.log
 
-Status = util.Enum('Origin', {
+Status = util.Enum('Status', {
     'ok':       'ok',       # Device's last seen time is shorter than expected interval for checkin, no errors listed.
     'missing':  'missing',  # Device's last seen time is longer than the expected interval for checkin, but no errors listed.
     'error':    'error' ,   # Device has errors listed.
     'unknown':  'unknown'   # Device did not set an expected checkin interval.
 })
-
-
-def require_login(handler_method):
-    """
-    A decorator to ensure the request is not a public request.
-
-    Accepts superuser and non-superuser requests.
-    Accepts drone and user requests.
-    """
-    def check_login(self, *args, **kwargs):
-        if self.public_request:
-            raise APIPermissionException('Login required.')
-        return handler_method(self, *args, **kwargs)
-    return check_login
-
-def require_superuser(handler_method):
-    """
-    A decorator to ensure the request is made as superuser.
-
-    Accepts drone and user requests.
-    """
-    def check_superuser(self, *args, **kwargs):
-        if not self.superuser_request:
-            raise APIPermissionException('Superuser required.')
-        return handler_method(self, *args, **kwargs)
-    return check_superuser
-
-def require_drone(handler_method):
-    """
-    A decorator to ensure the request is made as a drone.
-
-    Will also ensure superuser, which is implied with a drone request.
-    """
-    def check_drone(self, *args, **kwargs):
-        if self.origin.get('type', '') != Origin.device:
-            raise APIPermissionException('Drone request required.')
-        if not self.superuser_request:
-            raise APIPermissionException('Superuser required.')
-        return handler_method(self, *args, **kwargs)
-    return check_drone
 
 
 class DeviceHandler(base.RequestHandler):
@@ -79,10 +40,12 @@ class DeviceHandler(base.RequestHandler):
     def post(self):
         device_id = self.origin.get('id', '')
         payload = self.request.json_body
-        #validators.validate_data(payload, 'device.json', 'input', 'PUT', optional=True)
+        # Clean this up when validate_data method is fixed to use new schemas
+        # POST unnecessary, used to avoid run-time modification of schema
+        validate_data(payload, 'device.json', 'input', 'POST', optional=True)
 
         result = self.storage.update_el(device_id, payload)
-        if result.modified_count == 1:
+        if result.matched_count == 1:
             return {'modified': result.modified_count}
         else:
             raise APINotFoundException('Device with id {} not found, state not updated'.format(device_id))
